(355) Refactor Redux layer into actions/ and reducers/ folders.
