+++ src/index.js
@@ -1,10 +1,11 @@
 import React from 'react';
-import {Provider} from 'react-redux';
-import {store} from 'state/store';
+import { Provider } from 'react-redux';
+
 import Navigation from './routes';
+import { store } from './store';
+
 export default function App() {
   return (
-
     <Provider store={store}>
       <Navigation />
     </Provider>
