--- src/state/rootReducer.js
@@ -1,11 +1,0 @@
-import { combineReducers } from 'redux';
-
-import authReducer from './user/userReducer';
-import tripReducer from './trip/tripReducer';
-
-const rootReducer = combineReducers({
-  auth: authReducer,
-  trips: tripReducer,
-});
-
-export default rootReducer;
