--- src/state/trip/tripTypes.js
@@ -1,3 +1,0 @@
-export const DELETE_TRIP = 'DELETE_TRIP';
-export const CREATE_TRIP = 'CREATE_TRIP';
-export const SET_TRIPS = 'SET_TRIPS';
