--- src/state/user/userReducer.js
@@ -1,28 +1,0 @@
-import {LOGIN, SIGNUP, AUTHENTICATE} from './userActions';
-
-const initialState = {
-  token: null,
-  userId: null,
-};
-
-export default (state = initialState, action) => {
-  switch (action.type) {
-    case AUTHENTICATE:
-      return {
-        token: action.token,
-        userId: action.userId,
-      };
-    case LOGIN:
-      return {
-        token: action.token,
-        userId: action.userId,
-      };
-    case SIGNUP:
-      return {
-        token: action.token,
-        userId: action.userId,
-      };
-    default:
-      return state;
-  }
-};
