+++ src/store.js
@@ -1,0 +1,6 @@
+import { applyMiddleware, createStore } from 'redux';
+import thunk from 'redux-thunk';
+
+import rootReducer from 'reducers/rootReducer';
+
+export const store = createStore(rootReducer, applyMiddleware(thunk));
